Further differentiate between olivia and oracle
The `olivia::{Attestation, Announcement}` is used for deserializing
and modelling. The `oracle::Attestation` is just a wrapper which
allows us to send around attestations between different
`xtra::Actor`s.
