+++ daemon-tests/src/mocks/oracle.rs
@@ -1,5 +1,6 @@
 use crate::maia::OliviaData;
 use async_trait::async_trait;
+use daemon::olivia;
 use daemon::olivia::BitMexPriceEventId;
 use daemon::oracle;
 use mockall::*;
@@ -71,14 +72,14 @@ pub trait Oracle {
 /// announcement/attestation is hard-coded in OliviaData struct (along with event id's).
 /// Therefore, an attestation based on current utc time will always be wrong.
 pub fn dummy_wrong_attestation() -> oracle::Attestation {
-    let oracle::Attestation {
+    let olivia::Attestation {
         id: _,
         price,
         scalars,
-    } = OliviaData::example_0().attestation();
-    oracle::Attestation {
+    } = OliviaData::example_0().attestation().into_inner();
+    oracle::Attestation::new(olivia::Attestation {
         id: BitMexPriceEventId::with_20_digits(OffsetDateTime::now_utc()),
         price,
         scalars,
-    }
+    })
 }
