+++ daemon/src/taker_cfd.rs
@@ -145,8 +145,12 @@ impl<O, W> Actor<O, W> {
 #[xtra_productivity(message_impl = false)]
 impl<O, W> Actor<O, W> {
     async fn handle_attestation(&mut self, msg: oracle::Attestation) {
-        if let Err(e) =
-            cfd_actors::handle_oracle_attestation(msg, &self.db, &self.process_manager_actor).await
+        if let Err(e) = cfd_actors::handle_oracle_attestation(
+            msg.as_inner(),
+            &self.db,
+            &self.process_manager_actor,
+        )
+        .await
         {
             tracing::warn!("Failed to handle oracle attestation: {:#}", e)
         }
